pyproject.toml
[project]
name = "mcp-myguides"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.5",
    "pyyaml",
    "httpx",
    "pytest",
    "pydantic",
    "pytest-asyncio>=1.1.0",
    "pydantic-settings>=2.10.1",
]

[project.scripts]
mcp-server = "main:main"

[tool.pytest.ini_options]
pythonpath = "."


